service-workers/service-worker/fetch-filter-escalation.https.html

<!DOCTYPE html>
<title>Service Worker: fetch filter escalation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js?pipe=sub"></script>
<script>
var host_info = get_host_info();
var SCOPE = './resources/';
var SCRIPT = 'resources/fetch-filter-escalation-worker.js';

promise_test(function(t) {
    return service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(frame) {
          t.add_cleanup(function() { frame.remove(); });
          return frame.contentWindow.fetch(
              host_info.HTTPS_ORIGIN + '/service-workers/service-worker/resources/proxy');
        })
      .then(function(response) {
          assert_equals(response.type, 'basic');
          assert_true(response.headers.has('x-service-worker'));
        });
}, 'Local origin through Service Worker -- passes in Chromium (57.0.2987.98) -- passes in Firefox Nightly 55.0a1 (2017-04-11)');

promise_test(function(t) {
    return service_worker_unregister_and_register(t, SCRIPT, SCOPE)
      .then(function(registration) {
          return wait_for_state(t, registration.installing, 'activated');
        })
      .then(function() { return with_iframe(SCOPE); })
      .then(function(frame) {
          t.add_cleanup(function() { frame.remove(); });
          return frame.contentWindow.fetch(
              host_info.HTTPS_REMOTE_ORIGIN + '/service-workers/service-worker/resources/proxy')
        })
      .then(function(response) {
          assert_equals(response.type, 'basic');
          assert_true(response.headers.has('x-service-worker'));
        });
  }, 'Remote origin through Service Worker -- passes in Chromium (57.0.2987.98) -- fails in Firefox Nightly 55.0a1 (2017-04-11)');
</script>
